Made cart file. Filled search page. Made product pages.
Made the html file for the cart page. Filled the search page with a
list of all the possible results. Filled product html pages with
product details info. Added content to the main product category pages
to make it so that an image shows up beside the title in the product
list. Resized the images.